# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Affordly AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SimulationInput {
  /** @minLength 1 */
  scenarioName: string;
  monthlyIncome: number;
  monthlyRent: number;
  monthlyFood: number;
  monthlyTransport: number;
  monthlySubscriptions: number;
  monthlyOther: number;
  savingsBalance: number;
  debtBalance: number;
  decisionCost: number;
  /** @nullable */
  monthlyExtraCost?: number | null;
}

export type SimulationRiskLevel = typeof SimulationRiskLevel[keyof typeof SimulationRiskLevel];


export const SimulationRiskLevel = {
  Safe: 'Safe',
  Moderate: 'Moderate',
  Risky: 'Risky',
  Dangerous: 'Dangerous',
} as const;

export interface Simulation {
  id: number;
  scenarioName: string;
  monthlyIncome: number;
  monthlyRent: number;
  monthlyFood: number;
  monthlyTransport: number;
  monthlySubscriptions: number;
  monthlyOther: number;
  savingsBalance: number;
  debtBalance: number;
  decisionCost: number;
  /** @nullable */
  monthlyExtraCost?: number | null;
  totalMonthlyExpenses: number;
  monthlyCashflow: number;
  savingsRate: number;
  /** @nullable */
  monthsSavingsLast: number | null;
  affordabilityScore: number;
  riskLevel: SimulationRiskLevel;
  aiInsight: string;
  projectedSavings12m: number[];
  savingsAfterDecision?: number;
  recommendations?: string[];
  createdAt: string;
}

export interface DashboardSummary {
  totalSimulations: number;
  avgAffordabilityScore: number;
  safeCount: number;
  moderateCount: number;
  riskyCount: number;
  dangerousCount: number;
  avgMonthlyCashflow: number;
  avgSavingsRate: number;
}