# lib/api-zod/src/generated/types/simulation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Affordly AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SimulationRiskLevel } from './simulationRiskLevel';

export interface Simulation {
  id: number;
  scenarioName: string;
  monthlyIncome: number;
  monthlyRent: number;
  monthlyFood: number;
  monthlyTransport: number;
  monthlySubscriptions: number;
  monthlyOther: number;
  savingsBalance: number;
  debtBalance: number;
  decisionCost: number;
  /** @nullable */
  monthlyExtraCost?: number | null;
  totalMonthlyExpenses: number;
  monthlyCashflow: number;
  savingsRate: number;
  /** @nullable */
  monthsSavingsLast: number | null;
  affordabilityScore: number;
  riskLevel: SimulationRiskLevel;
  aiInsight: string;
  projectedSavings12m: number[];
  savingsAfterDecision?: number;
  recommendations?: string[];
  createdAt: string;
}